% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_team_schedule.R
\name{get_team_schedule}
\alias{get_team_schedule}
\title{get_team_schedule}
\usage{
get_team_schedule(team_id = NA_character_, team_url = NA_character_)
}
\arguments{
\item{team_id}{the team ID for a given team}

\item{team_url}{portion of team URL after /teams/}
}
\value{
a data frame with the following columns: date, game_id, home_team, home_team_id,
away_team, away_team_id, home_score, away_score, is_neutral, details
}
\description{
Scrapes the schedule for a given team from the NCAA. Returns data similar to
\code{bigballR} but does so more efficiently and quickly.
}
