% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaa_bb.R
\name{ncaa_bb_box_html}
\alias{ncaa_bb_box_html}
\title{ncaa_bb_box_html}
\usage{
ncaa_bb_box_html(game_id)
}
\arguments{
\item{game_id}{a unique identifier for an NCAA BB game}
}
\value{
an rvest html object
}
\description{
helper function to grab the raw html for a given game's box
}
