% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaa_bb.R
\name{ncaa_bb_calc_team_box_from_pbp}
\alias{ncaa_bb_calc_team_box_from_pbp}
\title{ncaa_bb_calc_team_box_from_pbp}
\usage{
ncaa_bb_calc_team_box_from_pbp(pbp)
}
\arguments{
\item{pbp}{a dataframe of pbp data created by \code{ncaa_bb_pbp}}
}
\value{
a dataframe formatted identically to the data returned by \code{ncaa_bb_team_box}
}
\description{
calculates team box score stats from pbp dataframe for ncaa
basketball games
}
