% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaa_bb_games.R
\name{ncaa_bb_games}
\alias{ncaa_bb_games}
\title{ncaa_bb_games}
\usage{
ncaa_bb_games(date = Sys.Date(), type = "men")
}
\arguments{
\item{date}{An R Date object. Defaults to \code{Sys.Date()}}

\item{type}{Men's or Women's game schedule (options: 'men', 'women')}
}
\value{
a data frame with NCAA MBB game data
}
\description{
Obtains basketball game data by date from the ncaa stats API. Some notes:
\itemize{
\item When values are missing (such as a team rank, or a bracket round), they are recorded as NA
\item dates and times are recorded in ET as is returned by the ET
\item If seeking to group by a team identifier, it is recommended to use slugs
\item game_id can be used to obtain pbp and box score data for the game
}
}
\examples{
\donttest{
  ncaa_bb_games(as.Date('2021-12-03'))
}
}
