% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaa_bb.R
\name{ncaa_bb_player_box}
\alias{ncaa_bb_player_box}
\title{ncaa_bb_player_box}
\usage{
ncaa_bb_player_box(game_id)
}
\arguments{
\item{game_id}{the game_id for a given game on stats.ncaa.org}
}
\value{
a dataframe for box score data for that game
}
\description{
Obtains player box score data for a given basketball game_id
from stats.ncaa.org.
}
