% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaa_bb.R
\name{ncaa_bb_player_box_parse}
\alias{ncaa_bb_player_box_parse}
\title{ncaa_bb_player_box_parse}
\usage{
ncaa_bb_player_box_parse(box_html)
}
\arguments{
\item{box_html}{an rvest html object}
}
\value{
a dataframe of cleaned, parsed player box data for the specified
\code{game_id}
}
\description{
a function to parse box html into a player box score table
}
