% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaa_bb_player_box_score.R
\name{ncaa_bb_player_box_score}
\alias{ncaa_bb_player_box_score}
\title{ncaa_bb_player_box_score}
\usage{
ncaa_bb_player_box_score(game_id)
}
\arguments{
\item{game_id}{the game_id of a given NCAA BB game from the NCAA API}
}
\value{
a data frame with NCAA BB player box score data for the given game
}
\description{
Obtains basketball game player box scores from the NCAA API.
}
\examples{
\donttest{
  ncaa_bb_player_box_score(5897924)
}
}
