% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaa_fb_box_score.R
\name{ncaa_fb_box_score}
\alias{ncaa_fb_box_score}
\title{ncaa_fb_box_score}
\usage{
ncaa_fb_box_score(game_id)
}
\arguments{
\item{game_id}{the game_id of a given NCAA FB game from the NCAA API}
}
\value{
a data frame with NCAA FB player box score data for the given game
}
\description{
Obtains football game team box scores from the NCAA API.
}
\examples{
\donttest{
  ncaa_fb_box_score(game_id = 5931773)
}
}
