% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaa_fb_game_info.R
\name{ncaa_fb_game_info}
\alias{ncaa_fb_game_info}
\title{ncaa_fb_game_info}
\usage{
ncaa_fb_game_info(game_id)
}
\arguments{
\item{game_id}{the game_id of a given NCAA FB game from the NCAA API}
}
\value{
a data frame with NCAA FB game info with a single row
}
\description{
Obtains football game team box scores from the NCAA API
}
\examples{
\donttest{
  ncaa_fb_game_info(5931773)
}
}
