% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaa_fb_games.R
\name{ncaa_fb_games}
\alias{ncaa_fb_games}
\title{ncaa_fb_games}
\usage{
ncaa_fb_games(week, year)
}
\arguments{
\item{week}{the week number of the given season}

\item{year}{the year for the season you wish to look at}
}
\value{
a data frame with all games for the given, week, year
}
\description{
Obtains a list of games for a given week and year of the season
}
\examples{
\donttest{
  ncaa_fb_games(10, 2021)
}
}
