% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaa_fb_pbp.R
\name{ncaa_fb_pbp}
\alias{ncaa_fb_pbp}
\title{ncaa_fb_pbp}
\usage{
ncaa_fb_pbp(game_id)
}
\arguments{
\item{game_id}{the game_id of a given NCAA FB game from the NCAA API}
}
\value{
a data frame of rows for each play in the game
}
\description{
Obtains the play by play info for a NCAA FB game
}
\examples{
\donttest{
  ncaa_fb_pbp(5931773)
}
}
