% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaa_fb_scoring_summary.R
\name{ncaa_fb_scoring_summary}
\alias{ncaa_fb_scoring_summary}
\title{ncaa_fb_scoring_summary}
\usage{
ncaa_fb_scoring_summary(game_id)
}
\arguments{
\item{game_id}{the game_id of a given NCAA FB game from the NCAA API}
}
\value{
A data frame of all scoring plays for the given game
}
\description{
Obtains football game scoring summary.
}
\examples{
\donttest{
  ncaa_fb_scoring_summary(5931773)
}
}
