% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaa_fb_team_stats.R
\name{ncaa_fb_team_stats}
\alias{ncaa_fb_team_stats}
\title{ncaa_fb_game_info}
\usage{
ncaa_fb_team_stats(game_id)
}
\arguments{
\item{game_id}{the game_id of a given NCAA FB game from the NCAA API}
}
\value{
a data frame with NCAA FB game team stats
}
\description{
Obtains football game team box scores from the NCAA API.
}
\examples{
\donttest{
  ncaa_fb_team_stats(5931773)
}
}
