% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaa_game_info.R
\name{ncaa_game_info}
\alias{ncaa_game_info}
\title{\strong{NCAA Game Info}}
\usage{
ncaa_game_info(game_id)
}
\arguments{
\item{game_id}{the game_id of a given game from the NCAA API}
}
\value{
A data frame of game information
}
\description{
\strong{NCAA Game Info}
}
