% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaa_roster.R
\name{ncaa_rosters}
\alias{ncaa_rosters}
\title{ncaa_rosters}
\usage{
ncaa_rosters(team_id)
}
\arguments{
\item{team_id}{a unique team identifier for a given NCAA team from any sport}
}
\value{
a dataframe of roster data with the \code{team}, \code{team_id}, \code{player},
\code{player_id}, \code{pos}, \code{year}, \code{games_played}, and \code{games_started} values for
each player on a given team. Where available, the \code{height} of players is
returned, in inches.
}
\description{
a generalized function for obtaining NCAA rosters for individual
teams
}
\examples{
\donttest{
ncaa_rosters('505669')
}

}
