% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaa_schedule.R
\name{ncaa_schedule}
\alias{ncaa_schedule}
\title{ncaa_schedule}
\usage{
ncaa_schedule(team_id)
}
\arguments{
\item{team_id}{the team ID for a given team}
}
\value{
a data frame with the following columns: date, game_id, home_team, home_team_id,
away_team, away_team_id, home_score, away_score, is_neutral, details
}
\description{
Scrapes the schedule for a given team from the NCAA. Returns data similar to
\code{bigballR} but does so more efficiently and quickly.
}
\examples{
\donttest{
ncaa_schedule('501404') # womens soccer
ncaa_schedule('509092') # baseball
ncaa_schedule('509675') # mens tennis
}

}
