% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaa_vb_adv_box_score.R
\name{ncaa_vb_adv_box_score}
\alias{ncaa_vb_adv_box_score}
\title{ncaa_vb_adv_box_score}
\usage{
ncaa_vb_adv_box_score(game_id, away = "Away", home = "Home")
}
\arguments{
\item{game_id}{the ID for the game's play_by_play data from the box score page}

\item{away}{Away Team}

\item{home}{Home Team}
}
\value{
a data frame with the following columns:
\itemize{
\item service point \%
\item serve \%
\item sideout \%
\item hurtful error \%
\item block win \%
\item forfeited points
\item total points generated
}

a data frame of play_by_play data for the specified game
}
\description{
Generates a simple advanced box score from volleyball play by play data.
}
\details{
How to get a game ID:
\enumerate{
\item find the game you want to view data for via stats.ncaa.org's scoreboard
\item open up its box score page
\item Click the play-by-play tab
\item Copy the string of numbers at the end of the URL from your browser's address bar.
}
}
