% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaa_vb_game_info.R
\name{ncaa_vb_game_info}
\alias{ncaa_vb_game_info}
\title{ncaa_vb_game_info}
\usage{
ncaa_vb_game_info(game_id)
}
\arguments{
\item{game_id}{the ID for the game's data}
}
\value{
a data frame of of a single row for info about the selected game
}
\description{
Scrape player box score during a volleyball game
}
\details{
How to get a game ID:
\enumerate{
\item find the game you want to view data for via ncaa.com's scoreboard
\item open up its box score page
\item Copy the string of numbers at the end of the URL from your browser's address bar.
OR
\item call the ncaa_vb_games() function
\item Select the game_id for the given game
}
}
