% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaa_vb_games.R
\name{ncaa_vb_games}
\alias{ncaa_vb_games}
\title{ncaa_vb_games}
\usage{
ncaa_vb_games(year, month, day)
}
\arguments{
\item{year}{Year, 4 digit format (YYYY)}

\item{month}{Month, 2 digit format (MM)}

\item{day}{Day, 2 digit format (DD)}
}
\value{
a data frame of games from the NCAA website
}
\description{
Pulls a list of volleyball games from the NCAA website for a given day
}
