% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaa_vb_roster.R
\name{ncaa_vb_roster}
\alias{ncaa_vb_roster}
\title{ncaa_vb_roster}
\usage{
ncaa_vb_roster(school_id, team_id)
}
\arguments{
\item{school_id}{the id for the school across all seasons}

\item{team_id}{the id for the roster for that school's season}
}
\value{
a data frame of the roster for the team
}
\description{
Scrape roster information from the stats.ncaa.org website
}
\details{
How to get a school_id and team_id:
\enumerate{
\item find the team you want to view data for on the stats.ncaa.org website
\item select the "roster" tab from the team's page
\item note the two sets of numbers in the URL
\item the first number in the URL is the school_id, the second number is the team_id
}
}
