% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaa_vb_team_box_score.R
\name{ncaa_vb_team_box_score}
\alias{ncaa_vb_team_box_score}
\title{ncaa_vb_team_box_score}
\usage{
ncaa_vb_team_box_score(game_id)
}
\arguments{
\item{game_id}{the ID for the game's box score data}
}
\value{
a data frame of team box score data for the specified game
}
\description{
Scrape team box score during a volleyball game from ncaa.com data
}
\details{
How to get a game ID:
\enumerate{
\item find the game you want to view data for via ncaa.com's scoreboard
\item open up its box score page
\item Copy the string of numbers at the end of the URL from your browser's address bar.
OR
\item call the ncaa_vb_games() function
\item Select the game_id for the given game
}
}
