% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaascrapR-package.R, R/utils.R
\docType{package}
\name{ncaascrapR-package}
\alias{ncaascrapR}
\alias{ncaascrapR-package}
\title{ncaascrapR: Functions to Access NCAA Sports Data}
\description{
A set of utilities for working with National Collegiate Athletic Association (NCAA) <https://www.ncaa.com> data. It is a scraping and aggregating interface for the Data and Stats <https://stats.ncaa.org/> portals of the NCAA.

A set of utilities for working with National Collegiate Athletic Association (NCAA) <https://www.ncaa.com> data. It is a scraping and aggregating interface for the Data and Stats <https://stats.ncaa.org/> portals of the NCAA.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ehess.github.io/ncaascrapR/}
  \item \url{https://github.com/ehess/ncaascrapR}
  \item Report bugs at \url{https://github.com/ehess/ncaascrapR/issues}
}


Useful links:
\itemize{
  \item \url{https://ehess.github.io/ncaascrapR/}
  \item \url{https://github.com/ehess/ncaascrapR}
  \item Report bugs at \url{https://github.com/ehess/ncaascrapR/issues}
}

}
\author{
\strong{Maintainer}: Eric Hess \email{eric.hess.13@gmail.com}

Authors:
\itemize{
  \item Akshay Easwaran \email{akeaswaran@me.com}
  \item John Edwards \email{eric.hess.13@gmail.com}
}

Other contributors:
\itemize{
  \item Saiem Gilani \email{saiem.gilani@gmail.com} [contributor]
}


\strong{Maintainer}: Eric Hess \email{eric.hess.13@gmail.com}

Authors:
\itemize{
  \item Akshay Easwaran \email{akeaswaran@me.com}
  \item John Edwards \email{eric.hess.13@gmail.com}
}

Other contributors:
\itemize{
  \item Saiem Gilani \email{saiem.gilani@gmail.com} [contributor]
}

}
\keyword{internal}
