% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_team_urls.R
\name{scrape_team_urls}
\alias{scrape_team_urls}
\title{scrape_team_urls}
\usage{
scrape_team_urls(year, sport_code, division = 1)
}
\arguments{
\item{year}{the year of the season}

\item{sport_code}{a code indicating which sport to scrape}

\item{division}{the division of nCAA competition. Defaults to 1, can be 2 or
3.}
}
\value{
a dataframe of team_ids and school names for the provided params.
}
\description{
scrapes team_urls for a given year, sport code, and division
}
