% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaa_vb_games.R
\name{ncaa_vb_games}
\alias{ncaa_vb_games}
\title{ncaa_vb_games}
\usage{
ncaa_vb_games(team_id)
}
\arguments{
\item{team_id}{the id for the team for that season}
}
\value{
a data frame of games from the NCAA website
}
\description{
Pulls a list of volleyball games from the NCAA website for a given day
}
