% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaa_vb_pbp.R
\name{ncaa_vb_pbp}
\alias{ncaa_vb_pbp}
\title{ncaa_vb_pbp}
\usage{
ncaa_vb_pbp(game_id)
}
\arguments{
\item{game_id}{the ID for the game's play_by_play data from the box score page}
}
\value{
a data frame of play-by-play data for the specified game
}
\description{
Scrapes plays during a volleyball game from the NCAA website.
}
\details{
How to get a game ID:
\enumerate{
\item find the game you want to view data for via stats.ncaa.org's scoreboard
\item open up its box score page
\item Click the play-by-play tab
\item Copy the string of numbers at the end of the URL from your browser's address bar.
}
}
